Make the code more modular
Move s3 to their module
